LPS-80392 Move impl to impl bundle
