+++ modules/apps/web-experience/application-list/application-list-impl/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly project(":apps:static:osgi:osgi-util")
+	compileOnly project(":apps:web-experience:application-list:application-list-api")
+	compileOnly project(":core:osgi-service-tracker-collections")
+}
